pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mantissa-stance"
version = "0.1.0"
description = "Open-source CSPM, DSPM, CIEM, and Vulnerability Management for AWS, GCP, and Azure"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
keywords = ["cspm", "dspm", "ciem", "cloud-security", "vulnerability-management", "aws", "gcp", "azure", "security", "multi-cloud", "iam", "posture-management"]

dependencies = [
    "boto3>=1.34.0",
]

[project.optional-dependencies]
# GCP support
gcp = [
    "google-cloud-storage>=2.14.0",
    "google-cloud-compute>=1.14.0",
    "google-cloud-iam>=2.12.0",
    "google-cloud-resource-manager>=1.10.0",
    "google-cloud-securitycenter>=1.23.0",
]

# Azure support
azure = [
    "azure-identity>=1.15.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-storage>=21.0.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-mgmt-network>=25.0.0",
    "azure-mgmt-security>=5.0.0",
    "azure-storage-blob>=12.19.0",
]

# All cloud providers
all = [
    "mantissa-stance[gcp,azure]",
]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "boto3-stubs[essential]>=1.34.0",
]

[project.scripts]
stance = "stance.cli:main"

[project.urls]
Homepage = "https://github.com/clay-good/mantissa-stance"
Repository = "https://github.com/clay-good/mantissa-stance"
Documentation = "https://github.com/clay-good/mantissa-stance/tree/main/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["stance"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["boto3.*", "botocore.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google.*", "google.cloud.*", "google.auth.*", "google.oauth2.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["azure.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/stance"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
